requirements.txt
openai 
streamlit 
streamlit-chat
python-dotenv
pytesseract 
pillow
pandas
python-docx